convert basic auth to FC and add hook
